Extract inline add row hook
