+++ vuu-ui/packages/vuu-table-extras/src/inline-add-row/index.ts
@@ -1,1 +1,2 @@
 export { InlineAddRow, type InlineAddRowProps } from "./InlineAddRow";
+export { useInlineAddRow, type UseInlineAddRowProps } from "./useInlineAddRow";
